Protobuf conversion: Use proto generated Path type
Fixes #1671
