+++ crates/dc_bundle/src/legacy_definition/element/path.rs
@@ -75,171 +75,3 @@ impl Default for Stroke {
         }
     }
 }
-
-// This is a simple vector representation that should reproduce Figma or
-// another design tool's rendering. It simplifies the renderer's task by
-// precomputing the stroke path, and using the same definitions as the
-// toolkit schema.
-//
-// The serialized form should be somewhat compact and fast to deserialize
-// because we're storing different types (commands, values) in big arrays,
-// reducing the number of objects that need to be constructed during
-// deserialization.
-
-#[repr(u8)]
-#[derive(Clone, Copy, PartialEq, Debug)]
-pub enum PathCommand {
-    MoveTo = 0,  // 1 Point
-    LineTo = 1,  // 1 Point
-    CubicTo = 2, // 3 Points
-    QuadTo = 3,  // 2 Points
-    Close = 4,   // 0 Points
-}
-impl TryFrom<u8> for PathCommand {
-    type Error = &'static str;
-
-    fn try_from(value: u8) -> Result<Self, Self::Error> {
-        match value {
-            0 => Ok(PathCommand::MoveTo),
-            1 => Ok(PathCommand::LineTo),
-            2 => Ok(PathCommand::CubicTo),
-            3 => Ok(PathCommand::QuadTo),
-            4 => Ok(PathCommand::Close),
-            _ => Err("PathCommand out of range"),
-        }
-    }
-}
-
-#[derive(Clone, Copy, PartialEq, Debug, Serialize, Deserialize, Hash)]
-#[serde(rename_all = "UPPERCASE")]
-pub enum WindingRule {
-    NonZero,
-    EvenOdd,
-    #[serde(other)]
-    None,
-}
-
-fn default_winding_rule() -> WindingRule {
-    WindingRule::None
-}
-
-#[derive(Clone, PartialEq, Serialize, Deserialize)]
-pub struct Path {
-    pub commands: Vec<u8>,
-    pub data: Vec<f32>,
-    #[serde(default = "default_winding_rule")]
-    pub winding_rule: WindingRule,
-}
-impl Path {
-    pub fn new() -> Path {
-        Path { commands: Vec::new(), data: Vec::new(), winding_rule: WindingRule::NonZero }
-    }
-    pub fn winding_rule(&mut self, winding_rule: WindingRule) -> &mut Path {
-        self.winding_rule = winding_rule;
-        self
-    }
-    pub fn move_to(&mut self, x: f32, y: f32) -> &mut Path {
-        self.commands.push(PathCommand::MoveTo as u8);
-        self.data.push(x);
-        self.data.push(y);
-        self
-    }
-    pub fn line_to(&mut self, x: f32, y: f32) -> &mut Path {
-        self.commands.push(PathCommand::LineTo as u8);
-        self.data.push(x);
-        self.data.push(y);
-        self
-    }
-    pub fn cubic_to(
-        &mut self,
-        c1_x: f32,
-        c1_y: f32,
-        c2_x: f32,
-        c2_y: f32,
-        x: f32,
-        y: f32,
-    ) -> &mut Path {
-        self.commands.push(PathCommand::CubicTo as u8);
-        self.data.push(c1_x);
-        self.data.push(c1_y);
-        self.data.push(c2_x);
-        self.data.push(c2_y);
-        self.data.push(x);
-        self.data.push(y);
-        self
-    }
-    pub fn quad_to(&mut self, c1_x: f32, c1_y: f32, x: f32, y: f32) -> &mut Path {
-        self.commands.push(PathCommand::QuadTo as u8);
-        self.data.push(c1_x);
-        self.data.push(c1_y);
-        self.data.push(x);
-        self.data.push(y);
-        self
-    }
-    pub fn close(&mut self) -> &mut Path {
-        self.commands.push(PathCommand::Close as u8);
-        self
-    }
-    pub fn iter(&self) -> PathIterator<'_> {
-        PathIterator(self.commands.iter(), self.data.iter())
-    }
-}
-
-/// A fused Command + Data type, making it easier to work with paths.
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub enum PathElement {
-    MoveTo { x: f32, y: f32 },
-    LineTo { x: f32, y: f32 },
-    CubicTo { c1_x: f32, c1_y: f32, c2_x: f32, c2_y: f32, x: f32, y: f32 },
-    QuadTo { c1_x: f32, c1_y: f32, x: f32, y: f32 },
-    Close,
-}
-
-pub struct PathIterator<'a>(std::slice::Iter<'a, u8>, std::slice::Iter<'a, f32>);
-
-/// Allow iteration over a path with commands and data fused.
-impl<'a> Iterator for PathIterator<'a> {
-    type Item = PathElement;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        match self.0.next().map(|val| PathCommand::try_from(*val)) {
-            Some(Ok(PathCommand::MoveTo)) => {
-                let x = *self.1.next()?;
-                let y = *self.1.next()?;
-                Some(PathElement::MoveTo { x, y })
-            }
-            Some(Ok(PathCommand::LineTo)) => {
-                let x = *self.1.next()?;
-                let y = *self.1.next()?;
-                Some(PathElement::LineTo { x, y })
-            }
-            Some(Ok(PathCommand::CubicTo)) => {
-                let c1_x = *self.1.next()?;
-                let c1_y = *self.1.next()?;
-                let c2_x = *self.1.next()?;
-                let c2_y = *self.1.next()?;
-                let x = *self.1.next()?;
-                let y = *self.1.next()?;
-                Some(PathElement::CubicTo { c1_x, c1_y, c2_x, c2_y, x, y })
-            }
-            Some(Ok(PathCommand::QuadTo)) => {
-                let c1_x = *self.1.next()?;
-                let c1_y = *self.1.next()?;
-                let x = *self.1.next()?;
-                let y = *self.1.next()?;
-                Some(PathElement::QuadTo { c1_x, c1_y, x, y })
-            }
-            Some(Ok(PathCommand::Close)) => Some(PathElement::Close),
-            _ => None,
-        }
-    }
-}
-
-impl std::fmt::Debug for Path {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("Path")
-            .field("winding_rule", &self.winding_rule)
-            .field("d", &self.iter().collect::<Vec<PathElement>>())
-            .finish()
-    }
-}
