+++ crates/dc_bundle/src/legacy_definition/element/view_shape.rs
@@ -14,7 +14,8 @@
  * limitations under the License.
  */
 use crate::definition::element::num_or_var::NumOrVar;
-use crate::legacy_definition::element::path::Path;
+//use crate::legacy_definition::element::path::Path;
+use crate::definition::element::Path;
 use serde::{Deserialize, Serialize};
 
 #[derive(Deserialize, Serialize, PartialEq, Debug, Clone)]
